floodnet/requirements.txt
zyndai-agent
langchain
langchain-google-genai
langchain-classic
langchain-community
langchain-core
httpx
requests
python-dotenv
